revert signer -> wallet crate renaming
