+++ crates/signer-aws/README.md
@@ -1,0 +1,5 @@
+# alloy-signer-aws
+
+Ethereum [AWS KMS] signer.
+
+[AWS KMS]: https://aws.amazon.com/kms
