+++ crates/signer-gcp/README.md
@@ -1,5 +1,5 @@
-# alloy-wallet-gcp
+# alloy-signer-gcp
 
-Ethereum [GCP KMS] wallet.
+Ethereum [GCP KMS] signer.
 
 [GCP KMS]: https://cloud.google.com/kms/docs/
