+++ crates/signer-ledger/README.md
@@ -1,0 +1,5 @@
+# alloy-signer-ledger
+
+Ethereum [Ledger] signer.
+
+[Ledger]: https://www.ledger.com
