+++ crates/signer-ledger/src/lib.rs
@@ -9,5 +9,8 @@
 #[macro_use]
 extern crate tracing;
 
-mod wallet;
-pub use wallet::{GcpKeyRingRef, GcpWallet, GcpWalletError, KeySpecifier};
+mod signer;
+pub use signer::LedgerWallet;
+
+mod types;
+pub use types::{DerivationType as HDPath, LedgerError};
