+++ crates/signer-ledger/src/signer.rs
@@ -110,7 +110,7 @@ impl LedgerWallet {
     ///
     /// ```
     /// # async fn foo() -> Result<(), Box<dyn std::error::Error>> {
-    /// use alloy_wallet_ledger::{HDPath, LedgerWallet};
+    /// use alloy_signer_ledger::{HDPath, LedgerWallet};
     ///
     /// let ledger = LedgerWallet::new(HDPath::LedgerLive(0), Some(1)).await?;
     /// # Ok(())
