+++ crates/signer-trezor/README.md
@@ -1,0 +1,5 @@
+# alloy-signer-trezor
+
+Ethereum [Trezor] signer.
+
+[Trezor]: https://trezor.io
