+++ crates/signer-trezor/src/lib.rs
@@ -9,8 +9,8 @@
 #[macro_use]
 extern crate tracing;
 
-mod wallet;
-pub use wallet::TrezorWallet;
+mod signer;
+pub use signer::TrezorWallet;
 
 mod types;
 pub use types::{DerivationType as HDPath, TrezorError};
