+++ crates/signer-trezor/src/signer.rs
