+++ crates/signer-trezor/src/types.rs
