--- crates/wallet-aws/README.md
@@ -1,5 +1,0 @@
-# alloy-wallet-aws
-
-Ethereum [AWS KMS] wallet.
-
-[AWS KMS]: https://aws.amazon.com/kms
