--- crates/wallet-ledger/README.md
@@ -1,5 +1,0 @@
-# alloy-wallet-ledger
-
-Ethereum [Ledger] wallet.
-
-[Ledger]: https://www.ledger.com
