--- crates/wallet-trezor/README.md
@@ -1,5 +1,0 @@
-# alloy-wallet-trezor
-
-Ethereum [Trezor] wallet.
-
-[Trezor]: https://trezor.io
